pyproject.toml
[project]
name = "sql2text"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "RameshArvind", email = "ramesharvind1994@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.3.3",
    "mcp>=1.0.0",
    "openai>=1.40.0",
    "linkup-sdk>=0.1.0",
    "deepl>=1.22.0",
]

[build-system]
requires = ["uv_build>=0.8.15,<0.9.0"]
build-backend = "uv_build"
